remove xarray_support module, use earthkit.data.readers.xarray
