--- src/anemoi/datasets/create/sources/xarray_support/README.md
@@ -1,1 +1,0 @@
-The code under this directory will be migrated to earthkit-data in the future
